move actions, reducers, and containers tests; update package.json to prevent tests from running endlessly
