+++ __tests__/containers/App.js
@@ -1,14 +1,14 @@
-jest.unmock('../src/js/containers/App.jsx')
-jest.unmock('../src/js/components/HomeLink.jsx')
-jest.mock('../src/js/redirect.js', () => {})
+jest.unmock('../../src/js/containers/App.jsx')
+jest.unmock('../../src/js/components/HomeLink.jsx')
+jest.mock('../../src/js/redirect.js', () => {})
 
 import React from 'react'
 import ReactDOM from 'react-dom'
 import TestUtils from 'react-addons-test-utils'
 import configureMockStore from 'redux-mock-store'
 import thunk from 'redux-thunk'
-import Wrapper from './Wrapper.js'
-import AppContainer from '../src/js/containers/App.jsx'
+import Wrapper from '../Wrapper.js'
+import AppContainer from '../../src/js/containers/App.jsx'
 
 const mockStore = configureMockStore([thunk])
